* commit '8a58f2b6eb07d875eaa2e0bfb88412cacfe1096e':
  fixed EXTZ-3389 & EXTZ-3394 issues
  fixed EXTZ-3389 & EXTZ-3394 issues
  fixed EXTZ-3389 & EXTZ-3394 issues

+++ grails-app/services/net/hedtech/banner/sspb/PageService.groovy
@@ -5,8 +5,11 @@
 package net.hedtech.banner.sspb
 
 import grails.converters.JSON
+import groovy.util.logging.Log4j
 import net.hedtech.banner.security.DeveloperSecurityService
+import net.hedtech.restfulapi.AccessDeniedException
 
+@Log4j
 class PageService {
     def compileService
     def groovyPagesTemplateEngine
@@ -110,6 +113,10 @@ class PageService {
     // TODO for now update(post) handles both update and creation to simplify client side logic
     def create(Map content, ignore) {
         log.trace "PageService.create invoked"
+        if (!developerSecurityService.isAllowModify(content.pageName, developerSecurityService.PAGE_IND)) {
+            log.error('user not authorized to create page')
+            throw new AccessDeniedException("user.not.authorized.create", [PBUser.getTrimmed().loginName])
+        }
         def result
         Page.withTransaction {
             // compile first
@@ -124,6 +131,10 @@ class PageService {
     // update is not used to update pages since the client may not know if a page exists or not when submitting (concurrent editing)
     def update( /*def id,*/ Map content, params) {
         log.trace "PageService.update invoked"
+        if (!developerSecurityService.isAllowModify(content.pageName, developerSecurityService.PAGE_IND)) {
+            log.error('user not authorized to update page')
+            throw new AccessDeniedException("user.not.authorized.update", [PBUser.getTrimmed().loginName])
+        }
         create(content, params)
     }
 
@@ -229,6 +240,10 @@ class PageService {
 
     // note the content-type header still needs to be set in the request even we don't send in any content in the body
     void delete(Map ignore, params) {
+        if (!developerSecurityService.isAllowModify(params.id, developerSecurityService.PAGE_IND)) {
+            log.error('user not authorized to delete page')
+            throw new AccessDeniedException("user.not.authorized.delete", [PBUser.getTrimmed().loginName])
+        }
         pageSecurityService.delete([:],[constantName:params.id])
         Page.withTransaction {
             def page = Page.find{constantName==params.id}
